+++ src/components/calcite-notice/calcite-notice.tsx
@@ -8,12 +8,6 @@ import {
   Method,
   Prop
 } from "@stencil/core";
-import {
-  lightbulb24F,
-  exclamationMarkTriangle24F,
-  checkCircle24F,
-  x32
-} from "@esri/calcite-ui-icons";
 import { getElementDir } from "../../utils/dom";
 
 /** Notices are intended to be used to present users with important-but-not-crucial contextual tips or copy. Because
@@ -95,6 +89,12 @@ export class CalciteNotice {
     if (!widths.includes(this.width)) this.width = "auto";
   }
 
+  componentDidLoad() {
+    this.noticeLinkEl = this.el.querySelectorAll(
+      "calcite-button"
+    )[0] as HTMLCalciteButtonElement;
+  }
+
   render() {
     const dir = getElementDir(this.el);
     const closeButton = (
@@ -102,22 +102,15 @@ export class CalciteNotice {
         class="notice-close"
         aria-label="close"
         onClick={() => this.close()}
+        ref={el => (this.closeButton = el)}
       >
-        <svg
-          xmlns="http://www.w3.org/2000/svg"
-          height="32"
-          width="32"
-          viewBox="0 0 32 32"
-        >
-          <path d={x32} />
-        </svg>
+        <calcite-icon icon="x" scale="s"></calcite-icon>
       </button>
     );
 
     return (
       <Host active={this.active} dir={dir}>
         {this.icon ? this.setIcon() : null}
-
         <div class="notice-content">
           <slot name="notice-title"></slot>
           <slot name="notice-message"></slot>
@@ -146,16 +139,28 @@ export class CalciteNotice {
   //
   //--------------------------------------------------------------------------
 
-  /** emit the `calciteNoticeClose` event - <calcite-notice> listens for this */
+  /** close the notice emit the `calciteNoticeClose` event - <calcite-notice> listens for this */
   @Method() async close() {
     this.active = false;
     this.calciteNoticeClose.emit({ requestedNotice: this.noticeId });
   }
 
-  /**  emit the `calciteNoticeOpen` event - <calcite-notice> listens for this  */
+  /** open the notice and emit the `calciteNoticeOpen` event - <calcite-notice> listens for this  */
   @Method() async open() {
     this.active = true;
     this.calciteNoticeOpen.emit({ requestedNotice: this.noticeId });
+  }
+
+  /** focus the close button, if present and requested */
+  @Method()
+  async setFocus() {
+    if (!this.closeButton && !this.noticeLinkEl) {
+      return;
+    }
+    if (this.noticeLinkEl) this.noticeLinkEl.setFocus();
+    else if (this.closeButton) {
+      this.closeButton.focus();
+    }
   }
 
   //--------------------------------------------------------------------------
@@ -167,25 +172,24 @@ export class CalciteNotice {
   /** Unique ID for this notice */
   private noticeId: string = this.el.id;
 
+  /** the close button element */
+  private closeButton?: HTMLElement;
+
+  /** the notice link child element  */
+  private noticeLinkEl?: HTMLCalciteButtonElement;
+
   private iconDefaults = {
-    green: checkCircle24F,
-    yellow: exclamationMarkTriangle24F,
-    red: exclamationMarkTriangle24F,
-    blue: lightbulb24F
+    green: "checkCircle",
+    yellow: "exclamationMarkTriangle",
+    red: "exclamationMarkTriangle",
+    blue: "lightbulb"
   };
 
   private setIcon() {
     var path = this.iconDefaults[this.color];
     return (
       <div class="notice-icon">
-        <svg
-          xmlns="http://www.w3.org/2000/svg"
-          height="24"
-          width="24"
-          viewBox="0 0 24 24"
-        >
-          <path d={path} />
-        </svg>
+        <calcite-icon icon={path} filled scale="s"></calcite-icon>
       </div>
     );
   }